Port typenetwork profile
